Clean up LINE-2D's DRAW-OBJECT method
This is close to as good as we can get performance and readability
wise, as far as I can tell. The crux of the issue is we can't get
around listing the functions and arguments to pick from without
sacrificing performance. Any attempt to dynamically build the args is
slower than just picking from a known execution path.