Defer parameters in make-scene-pro (needs cleanup)
This commit is a first properly working version of make-scene-pro, but
it needs some elegance injected.

Rationale: We sometimes need to defer the initialization of
parameters, such is the case with cameras because they require a
window to be initialized. I have temporarily made DEFER-INIT effect
parameters the same way as objects, but we may want more fine grained
control, perhaps at the level of individual scene items.

We also need to make sure that scene objects that refer to a parameter
are actually using that exact parameter and not a separate evaluation
of an identical form. This fix was made in the previous commit
(3581387). I believe it is still an issue with the regular MAKE-SCENE,
though it has slipped under the radar because evaluation of an
identical form often yields inconsequential differences. I noticed it
because billboards require a camera and 2 make-camera calls gave a
different camera to the scene and to the billboard.