No, we want to reuse the new Mesh data as long as it isn't freed
i.e. If a user makes an rl-mesh and SETFs an element of a model's
meshes array to the new rl-mesh, the user must watch out for freeing
the data that the rl-mesh's C pointer fields are pointing to while
they are in use. I think this won't be an issue. In Lisp thinking, you
wouldn't free the rl-mesh after setting it; one, you wouldn't normally
free anything; two, you wouldn't know that SETF is actually copying C
data under the hood. So, this should work nicely.